+++ src/cpm/mod.rs
@@ -36,7 +36,7 @@ mod machine;
 pub mod uart;
 
 pub use fcb::{parse_afn, parse_command_fcb, parse_dir_operand, split_8_3, Fcb, FCB_SIZE};
-pub use fs::CpmFs;
+pub use fs::{CpmFs, DEFAULT_DMA};
 pub use machine::CpmMachine;
 pub use uart::{resolve_access, ModemAccess};